+++ src/arche/rules/category.py
@@ -14,31 +14,35 @@ def get_difference(
     result = Result("Category Coverage Difference")
 
     for c in category_names:
-        cats = pd.DataFrame(
-            {
-                "s": source_df[c].value_counts(normalize=True, dropna=False),
-                "t": target_df[c].value_counts(normalize=True, dropna=False),
-            }
-        ).fillna(0)
-        cat_difs = (
-            ((cats["s"] - cats["t"]) * 100)
-            .abs()
-            .sort_index(kind="mergesort")
-            .sort_values()
-            .round(decimals=2)
+        cats = (
+            pd.DataFrame(
+                {
+                    source_key: source_df[c]
+                    .value_counts(dropna=False, normalize=True)
+                    .mul(100)
+                    .round(decimals=2),
+                    target_key: target_df[c]
+                    .value_counts(dropna=False, normalize=True)
+                    .mul(100)
+                    .round(decimals=2),
+                }
+            )
+            .fillna(0)
+            .sort_values(by=[source_key, target_key], kind="mergesort")
         )
+        cats.name = f"Coverage difference in {c}"
+        cat_difs = ((cats[source_key] - cats[target_key])).abs()
         cat_difs.name = (
             f"Coverage difference between {source_key}'s and {target_key}'s {c}"
         )
-        cat_difs = cat_difs[cat_difs > 0]
-        if cat_difs.empty:
-            continue
-        result.add_info(f"'{c}' PASSED", stats=cat_difs)
         errs = cat_difs[cat_difs > 20]
         if not errs.empty:
             result.add_warning(
                 f"The difference is greater than 20% for {len(errs)} value(s) of {c}"
             )
+
+        result.add_info(f"'{c}' PASSED", stats=cats)
+
     return result