+++ CHANGES.md
@@ -18,7 +18,7 @@ Keep a Changelog](https://keepachangelog.com/en/1.0.0/), [Semantic Versioning](h
 - *Fields Coverage* now is printed as a bar plot, #9
 - *Fields Counts* renamed to *Coverage Difference* and is printed as a bar plot, #9
 - *Compare Scraped Categories* renamed to *Category Coverage Difference*
-- Refactored *Category Coverage Difference* to return difference in categories coverage in percentage, #9
+- Refactored *Category Coverage Difference* to return categories coverage, #9
 - *Coverage Difference* finds the absolute difference in coverage in percentage (was difference in ratio before)
 - Bar plots xaxis changed from numeric to logarithmic